+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Handling of `PlanetBuild` submissions
+
 ### Changed
 
 - Apps now start with the package.json type set to module